backend/requirements.txt
Flask==2.3.3
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-telegram-bot==13.15
apscheduler==3.10.4